Updated style
I'm not good at reading code so I had to change a bunch of things. None of this changes how the code does anything, just how it looks.

Also I'm realizing that I should have done a branch before this. I will fix anything I broke on your end ASAP.

The other thing I did was remove the files that looked like they were for your class project to make it easier to focus.

Because I wanted to keep this as just the style changes, this does not contain my server code.